roll back splunk uf version and remove pcaps from source control

+++ scripts/get_pcap.py
@@ -5,7 +5,8 @@ import requests
 # TODO: pull the bucket/object values from a config file
 
 print("Renaming existing zeek_streamer.pcap file")
-os.rename('./pcaps/zeek_streamer.pcap', './pcaps/zeek_streamer.pcap.bak')
+if os.path.exists('./pcaps/zeek_streamer.pcap'):
+    os.rename('./pcaps/zeek_streamer.pcap', './pcaps/zeek_streamer.pcap.bak')
 
 url = 'https://bhayes-pcaps.s3.us-east-2.amazonaws.com/garage-2020-10-18.pcap'
 headers = {'Host': 'bhayes-pcaps.s3.us-east-2.amazonaws.com'}
